REMOVE: Removed Board.tsx and moved its content to BattlePage.tsx
