+++ src/app.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { BattlePage } from "./components/BattlePage";
+import  BattlePage  from "./components/BattlePage";
 import { WelcomePage } from "./components/WelcomePage";
 import { Layout } from "./components/layout/Layout";
 import { ConfigurationPanel } from "./components/ConfigurationPanel";
